+++ src/renderer/src/hooks/useUpdateHandler.ts
@@ -31,8 +31,8 @@ export default function useUpdateHandler() {
           title: t('button.update_available'),
           message: t('button.update_available', { version: releaseInfo.version }),
           silent: true,
-          channel: 'system',
-          timestamp: Date.now()
+          timestamp: Date.now(),
+          source: 'update'
         })
         dispatch(
           setUpdateState({
