+++ src/renderer/src/utils/window.ts
@@ -1,0 +1,3 @@
+export const isFocused = () => {
+  return document.hasFocus()
+}
